+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add error on invalid currency
 - Fix an error on tip to lowercase user registered
 - Fix an error on pending tips
+- Fix an error on withdraw in history
 
 # v1.1 - 16/06/2017
 - Add support of new syntax for tip command #57